requirements.txt
torch>=2.0.0
transformers>=4.0.0
diffusers>=0.9.0
Pillow>=9.0.0
numpy<2.0.0
accelerate~=0.31.0
matplotlib~=3.9.0
pandas~=2.2.2
requests